+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.5.0] 2020-04-21
+### Added
+- Support for configurable extensions and plugins;
+  - `Sequel.extension(:deferrable_foreign_keys)` now supports `by_default` option to support the
+    global default `deferrable:` value (`true` by default);
+
 ## [0.4.0] 2019-11-18
 ### Added
 - `Sequel.extension(:deferrable_foreign_keys)` - makes foreign keys constraints deferrable by default;